feat: implement GuideList component and refactor wallet connection UI for better user experience
