+++ src/Ver2Designs/Pages/Landing/LandingV3/HeroSection/styles.ts
@@ -8,7 +8,7 @@ export const heroSectionContainer: SxProps<Theme> = {
 export const heroSectionContentArea: SxProps<Theme> = {
   paddingTop: "80px",
   height: {
-    xs: "65dvh",
+    xs: "75dvh",
     sm: "70dvh",
     md: "70dvh",
     lg: "80dvh",
@@ -92,7 +92,7 @@ export const howItWorksButton: SxProps<Theme> = {
   textTransform: "capitalize",
   border: "none",
   borderRadius: "8px",
-  padding: "1rem 1.5rem",
+  padding: ".5rem 1.5rem",
   "&:hover": {
     backgroundColor: "#fff",
   },
@@ -105,7 +105,7 @@ export const getStartedButton: SxProps<Theme> = {
   backgroundColor: "#5265FF",
   textTransform: "capitalize",
   borderRadius: "8px",
-  padding: "1rem 1.5rem",
+  padding: ".5rem 1.5rem",
   color: "#fff",
   "&:hover": {
     backgroundColor: "#5265FF",
